Run process copy within process.
